Split Answers classes to a separate file
